refactor: email & password regex -> 상수화 & runCatching -> fold 적용
