+++ app/src/main/java/org/sopt/sample/presentation/signup/viewmodel/SignUpViewModel.kt
@@ -3,15 +3,13 @@ package org.sopt.sample.presentation.signup.viewmodel
 import androidx.lifecycle.*
 import kotlinx.coroutines.launch
 import org.sopt.sample.data.model.SignUpRequest
-import org.sopt.sample.data.repository.AuthRepository
+import org.sopt.sample.domain.repository.AuthRepository
 import org.sopt.sample.util.Event
 import org.sopt.sample.util.extension.addSourceList
 
 class SignUpViewModel(private val authRepository: AuthRepository) : ViewModel() {
-    private val emailRegex =
-        Regex("^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{6,10}+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+\$")
-    private val passwordRegex =
-        Regex("^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@\$!%*#?&])[A-Za-z\\d@\$!%*#?&]{6,12}\$")
+    private val emailRegex = Regex(EMAIL_REGEX)
+    private val passwordRegex = Regex(PASSWORD_REGEX)
 
     val email = MutableLiveData<String>()
     val isValidEmail: LiveData<Boolean> = Transformations.map(email) {
@@ -47,15 +45,22 @@ class SignUpViewModel(private val authRepository: AuthRepository) : ViewModel() 
         return isValidEmail.value!! && isValidPassword.value!! && isValidName.value!!
     }
 
-    fun signUp(signUpRequest: SignUpRequest) {
+    fun signUp() {
         viewModelScope.launch {
-            kotlin.runCatching {
-                authRepository.signUp(signUpRequest)
-            }.onSuccess {
+            runCatching {
+                authRepository.signUp(SignUpRequest(email.value!!, password.value!!, name.value!!))
+            }.fold({
                 _signUpEvent.value = Event(true)
-            }.onFailure {
+            }, {
                 _signUpEvent.value = Event(false)
-            }
+            })
         }
+    }
+
+    companion object {
+        private const val EMAIL_REGEX =
+            "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{6,10}+@[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+\$"
+        private const val PASSWORD_REGEX =
+            "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@\$!%*#?&])[A-Za-z\\d@\$!%*#?&]{6,12}\$"
     }
 }
